Port Linux/macOS to 64-bit, using new CMake build system.
Huge shoutout to dhewm3 and RBDOOM-3-BFG for doing 99% of this work before us!

+++ d3xp/Pvs.cpp
@@ -329,7 +329,9 @@ pvsStack_t *idPVS::FloodPassagePVS_r( pvsPortal_t *source, const pvsPortal_t *po
 	pvsArea_t *area;
 	pvsStack_t *stack;
 	pvsPassage_t *passage;
-	long *sourceVis, *passageVis, *portalVis, *mightSee, *prevMightSee, more;
+	// RB: 64 bit fixes, changed long to int
+	int *sourceVis, *passageVis, *portalVis, *mightSee, *prevMightSee, more;
+	// RB end
 
 	area = &pvsAreas[portal->areaNum];
 
@@ -364,15 +366,20 @@ pvsStack_t *idPVS::FloodPassagePVS_r( pvsPortal_t *source, const pvsPortal_t *po
 		source->vis[n >> 3] |= (1 << (n & 7));
 
 		// get pointers to vis data
-		prevMightSee = reinterpret_cast<long *>(prevStack->mightSee);
-		passageVis = reinterpret_cast<long *>(passage->canSee);
-		sourceVis = reinterpret_cast<long *>(source->vis);
-		mightSee = reinterpret_cast<long *>(stack->mightSee);
+
+		// RB: 64 bit fixes, changed long to int
+		prevMightSee = reinterpret_cast<int *>(prevStack->mightSee);
+		passageVis = reinterpret_cast<int *>(passage->canSee);
+		sourceVis = reinterpret_cast<int *>(source->vis);
+		mightSee = reinterpret_cast<int *>(stack->mightSee);
+		// RB end
 
 		more = 0;
 		// use the portal PVS if it has been calculated
 		if ( p->done ) {
-			portalVis = reinterpret_cast<long *>(p->vis);
+			// RB: 64 bit fixes, changed long to int
+			portalVis = reinterpret_cast<int *>(p->vis);
+			// RB end
 			for ( j = 0; j < portalVisLongs; j++ ) {
 				// get new PVS which is decreased by going through this passage
 				m = *prevMightSee++ & *passageVis++ & *portalVis++;
@@ -730,7 +737,9 @@ idPVS::AreaPVSFromPortalPVS
 */
 int idPVS::AreaPVSFromPortalPVS( void ) const {
 	int i, j, k, areaNum, totalVisibleAreas;
-	long *p1, *p2;
+	// RB: 64 bit fixes, changed long to int
+	int *p1, *p2;
+	// RB end
 	byte *pvs, *portalPVS;
 	pvsArea_t *area;
 
@@ -755,8 +764,10 @@ int idPVS::AreaPVSFromPortalPVS( void ) const {
 
 		// store the PVS of all portals in this area at the first portal
 		for ( j = 1; j < area->numPortals; j++ ) {
-			p1 = reinterpret_cast<long *>(area->portals[0]->vis);
-			p2 = reinterpret_cast<long *>(area->portals[j]->vis);
+			// RB: 64 bit fixes, changed long to int
+			p1 = reinterpret_cast<int *>(area->portals[0]->vis);
+			p2 = reinterpret_cast<int *>(area->portals[j]->vis);
+			// RB end
 			for ( k = 0; k < portalVisLongs; k++ ) {
 				*p1++ |= *p2++;
 			}
@@ -807,7 +818,9 @@ void idPVS::Init( void ) {
 	areaQueue = new int[numAreas];
 
 	areaVisBytes = ( ((numAreas+31)&~31) >> 3);
-	areaVisLongs = areaVisBytes/sizeof(long);
+	// RB: 64 bit fixes, changed long to int
+	areaVisLongs = areaVisBytes/sizeof(int);
+	// RB end
 
 	areaPVS = new byte[numAreas * areaVisBytes];
 	memset( areaPVS, 0xFF, numAreas * areaVisBytes );
@@ -815,7 +828,9 @@ void idPVS::Init( void ) {
 	numPortals = GetPortalCount();
 
 	portalVisBytes = ( ((numPortals+31)&~31) >> 3);
-	portalVisLongs = portalVisBytes/sizeof(long);
+	// RB: 64 bit fixes, changed long to int
+	portalVisLongs = portalVisBytes/sizeof(int);
+	// RB end
 
 	for ( int i = 0; i < MAX_CURRENT_PVS; i++ ) {
 		currentPVS[i].handle.i = -1;
@@ -1013,7 +1028,9 @@ idPVS::SetupCurrentPVS
 pvsHandle_t idPVS::SetupCurrentPVS( const int *sourceAreas, const int numSourceAreas, const pvsType_t type ) const {
 	int i, j;
 	unsigned int h;
-	long *vis, *pvs;
+	// RB: 64 bit fixes, changed long to int
+	int *vis, *pvs;
+	// RB end
 	pvsHandle_t handle;
 
 	h = 0;
@@ -1034,8 +1051,10 @@ pvsHandle_t idPVS::SetupCurrentPVS( const int *sourceAreas, const int numSourceA
 
 			assert( sourceAreas[i] >= 0 && sourceAreas[i] < numAreas );
 
-			vis = reinterpret_cast<long*>(areaPVS + sourceAreas[i] * areaVisBytes);
-			pvs = reinterpret_cast<long*>(currentPVS[handle.i].pvs);
+			// RB: 64 bit fixes, changed long to int
+			vis = reinterpret_cast<int*>(areaPVS + sourceAreas[i] * areaVisBytes);
+			pvs = reinterpret_cast<int*>(currentPVS[handle.i].pvs);
+			// RB end
 			for ( j = 0; j < areaVisLongs; j++ ) {
 				*pvs++ |= *vis++;
 			}
@@ -1074,7 +1093,9 @@ idPVS::MergeCurrentPVS
 */
 pvsHandle_t idPVS::MergeCurrentPVS( pvsHandle_t pvs1, pvsHandle_t pvs2 ) const {
 	int i;
-	long *pvs1Ptr, *pvs2Ptr, *ptr;
+	// RB: 64 bit fixes, changed long to int
+	int *pvs1Ptr, *pvs2Ptr, *ptr;
+	// RB end
 	pvsHandle_t handle;
 
 	if ( pvs1.i < 0 || pvs1.i >= MAX_CURRENT_PVS || pvs1.h != currentPVS[pvs1.i].handle.h ||
@@ -1084,9 +1105,11 @@ pvsHandle_t idPVS::MergeCurrentPVS( pvsHandle_t pvs1, pvsHandle_t pvs2 ) const {
 
 	handle = AllocCurrentPVS( pvs1.h ^ pvs2.h );
 
-	ptr = reinterpret_cast<long*>(currentPVS[handle.i].pvs);
-	pvs1Ptr = reinterpret_cast<long*>(currentPVS[pvs1.i].pvs);
-	pvs2Ptr = reinterpret_cast<long*>(currentPVS[pvs2.i].pvs);
+	// RB: 64 bit fixes, changed long to int
+	ptr = reinterpret_cast<int*>(currentPVS[handle.i].pvs);
+	pvs1Ptr = reinterpret_cast<int*>(currentPVS[pvs1.i].pvs);
+	pvs2Ptr = reinterpret_cast<int*>(currentPVS[pvs2.i].pvs);
+	// RB end
 
 	for ( i = 0; i < areaVisLongs; i++ ) {
 		*ptr++ = *pvs1Ptr++ | *pvs2Ptr++;
