+++ d3xp/ai/AAS_routing.cpp
@@ -167,7 +167,9 @@ void idAASLocal::CalculateAreaTravelTimes(void) {
 		}
 	}
 
-	assert( ( (unsigned int) bytePtr - (unsigned int) areaTravelTimes ) <= numAreaTravelTimes * sizeof( unsigned short ) );
+	// RB: 64 bit fixes, changed unsigned int to ptrdiff_t
+	assert( ( (ptrdiff_t) bytePtr - (ptrdiff_t) areaTravelTimes ) <= numAreaTravelTimes * sizeof( unsigned short ) );
+	// RB end
 }
 
 /*
